# openapi/auth-profiles.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fieldnote Care Auth and Profiles API",
    "version": "0.1.0",
    "description": "Generated from the FastAPI application. See API_CONVENTIONS.md for shared request, response, cookie, and CSRF conventions."
  },
  "paths": {
...
    "/api/v1/profiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List Profiles",
        "operationId": "list_profiles_api_v1_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  },
                  "type": "array",
                  "title": "Response List Profiles Api V1 Profiles Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Create Dependent Profile",
        "operationId": "create_dependent_profile_api_v1_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DependentProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/{profile_id}": {
      "patch": {
        "tags": [
          "profiles"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_v1_profiles__profile_id__patch",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_v1_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DependentProfileCreate": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "default": "UTC"
          },
          "relationship": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Relationship",
            "default": "dependent"
          },
          "avatar_color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Avatar Color",
            "default": "#456c58"
          },
          "avatar_wash": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Avatar Wash",
            "default": "#e5ede5"
          },
          "blood_group": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Group"
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Allergies"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Conditions"
          },
          "onboarding_step": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Onboarding Step",
            "default": 0
          },
          "onboarding_complete": {
            "type": "boolean",
            "title": "Onboarding Complete",
            "default": false
          },
          "family_id": {
            "type": "string",
            "format": "uuid",
            "title": "Family Id"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "family_id"
        ],
        "title": "DependentProfileCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          },
          "avatar_color": {
            "type": "string",
            "title": "Avatar Color"
          },
          "avatar_wash": {
            "type": "string",
            "title": "Avatar Wash"
          },
          "blood_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Group"
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Conditions"
          },
          "onboarding_step": {
            "type": "integer",
            "title": "Onboarding Step"
          },
          "onboarding_complete": {
            "type": "boolean",
            "title": "Onboarding Complete"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "display_name",
          "date_of_birth",
          "sex",
          "timezone",
          "relationship",
          "avatar_color",
          "avatar_wash",
          "blood_group",
          "allergies",
          "conditions",
          "onboarding_step",
          "onboarding_complete"
        ],
        "title": "ProfileResponse"
      },
      "ProfileUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "relationship": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          },
          "avatar_color": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Color"
          },
          "avatar_wash": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Wash"
          },
          "blood_group": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Group"
          },
          "allergies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Allergies"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "onboarding_step": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Step"
          },
          "onboarding_complete": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Complete"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}